Adds dependencies needed to set up the SQLite database and some slight refactoring where activities are placed in one package and fragments in another package
