+++ GreenGarbageApp/app/src/main/java/com/example/greengarbageapp/activities/MainActivity.kt
@@ -1,14 +1,12 @@
 package com.example.greengarbageapp.activities
 
 import android.os.Bundle
-import android.widget.TextView
 
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import androidx.drawerlayout.widget.DrawerLayout
 import androidx.navigation.NavController
 import androidx.navigation.NavDestination
-import androidx.navigation.findNavController
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.NavigationUI
