+++ GreenGarbageApp/app/src/main/java/com/example/greengarbageapp/fragments/EndFragment.kt
@@ -1,4 +1,4 @@
-package com.example.greengarbageapp.activities
+package com.example.greengarbageapp.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
