+++ GreenGarbageApp/app/src/main/java/com/example/greengarbageapp/fragments/LegalFragment.kt
@@ -1,11 +1,10 @@
-package com.example.greengarbageapp.activities
+package com.example.greengarbageapp.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.greengarbageapp.databinding.FragmentAboutBinding
 import com.example.greengarbageapp.databinding.FragmentLegalBinding
 
 class LegalFragment : Fragment() {
