+++ GreenGarbageApp/app/src/main/java/com/example/greengarbageapp/fragments/UNinfoFragment.kt
@@ -1,4 +1,4 @@
-package com.example.greengarbageapp.activities
+package com.example.greengarbageapp.fragments
 
 import android.content.Intent
 import android.net.Uri
@@ -7,7 +7,6 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.greengarbageapp.databinding.FragmentAboutBinding
 import com.example.greengarbageapp.databinding.FragmentUNinfoBinding
 
 class UNinfoFragment : Fragment() {
