+++ GreenGarbageApp/app/src/main/res/navigation/navigation.xml
@@ -6,7 +6,7 @@
     app:startDestination="@id/startFragment">
     <fragment
         android:id="@+id/startFragment"
-        android:name="com.example.greengarbageapp.activities.StartFragment"
+        android:name="com.example.greengarbageapp.fragments.StartFragment"
         android:label="Home">
         <action
             android:id="@+id/action_startFragment_to_introFragment"
@@ -14,7 +14,7 @@
     </fragment>
     <fragment
         android:id="@+id/introFragment"
-        android:name="com.example.greengarbageapp.activities.IntroFragment"
+        android:name="com.example.greengarbageapp.fragments.IntroFragment"
         android:label="Introduction">
         <action
             android:id="@+id/action_introFragment_to_startFragment"
@@ -25,7 +25,7 @@
     </fragment>
     <fragment
         android:id="@+id/gameFragment"
-        android:name="com.example.greengarbageapp.activities.GameFragment"
+        android:name="com.example.greengarbageapp.fragments.GameFragment"
         android:label="Game">
         <action
             android:id="@+id/action_gameFragment_to_introFragment"
@@ -33,18 +33,18 @@
     </fragment>
     <fragment
         android:id="@+id/aboutFragment"
-        android:name="com.example.greengarbageapp.activities.AboutFragment"
+        android:name="com.example.greengarbageapp.fragments.AboutFragment"
         android:label="About us" />
     <fragment
         android:id="@+id/UNinfoFragment"
-        android:name="com.example.greengarbageapp.activities.UNinfoFragment"
+        android:name="com.example.greengarbageapp.fragments.UNinfoFragment"
         android:label="UNs SDGS Goal 12" />
     <fragment
         android:id="@+id/legalFragment"
-        android:name="com.example.greengarbageapp.activities.LegalFragment"
+        android:name="com.example.greengarbageapp.fragments.LegalFragment"
         android:label="Legal" />
     <fragment
         android:id="@+id/contactFragment"
-        android:name="com.example.greengarbageapp.activities.ContactFragment"
+        android:name="com.example.greengarbageapp.fragments.ContactFragment"
         android:label="Contact us" />
 </navigation>
